requirements.txt
# Core dependencies
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
slowapi==0.1.8
pydantic==2.6.4

# GPU & model dependencies (optimized for unsloth)
torch==2.4.0
transformers==4.55.0
accelerate==0.34.1
unsloth==2025.8.4
sentence-transformers==2.7.0

# RAG & text processing
langchain==0.1.16
langchain-community==0.0.35
chromadb==1.0.16
pyvi==0.1.1

# DeepSeek API
openai==1.25.0

# Additional utilities
httpx==0.27.0
python-multipart==0.0.9